Fix PWA title bar theme color active vs inactive
