+++ src/vs/workbench/browser/style.ts
@@ -5,12 +5,10 @@
 
 import './media/style.css';
 import { registerThemingParticipant } from '../../platform/theme/common/themeService.js';
-import { WORKBENCH_BACKGROUND, TITLE_BAR_ACTIVE_BACKGROUND } from '../common/theme.js';
-import { isWeb, isIOS } from '../../base/common/platform.js';
-import { createMetaElement } from '../../base/browser/dom.js';
+import { WORKBENCH_BACKGROUND } from '../common/theme.js';
+import { isIOS } from '../../base/common/platform.js';
 import { isSafari, isStandalone } from '../../base/browser/browser.js';
 import { selectionBackground } from '../../platform/theme/common/colorRegistry.js';
-import { mainWindow } from '../../base/browser/window.js';
 
 registerThemingParticipant((theme, collector) => {
 
@@ -22,22 +20,6 @@ registerThemingParticipant((theme, collector) => {
 	const windowSelectionBackground = theme.getColor(selectionBackground);
 	if (windowSelectionBackground) {
 		collector.addRule(`.monaco-workbench ::selection { background-color: ${windowSelectionBackground}; }`);
-	}
-
-	// Update <meta name="theme-color" content=""> based on selected theme
-	if (isWeb) {
-		const titleBackground = theme.getColor(TITLE_BAR_ACTIVE_BACKGROUND);
-		if (titleBackground) {
-			const metaElementId = 'monaco-workbench-meta-theme-color';
-			let metaElement = mainWindow.document.getElementById(metaElementId) as HTMLMetaElement | null;
-			if (!metaElement) {
-				metaElement = createMetaElement();
-				metaElement.name = 'theme-color';
-				metaElement.id = metaElementId;
-			}
-
-			metaElement.content = titleBackground.toString();
-		}
 	}
 
 	// We disable user select on the root element, however on Safari this seems
